+++ src/router/index.ts
@@ -1,9 +1,9 @@
 import Vue from "vue";
 import Router from "vue-router";
 import PdfViewer from "@/views/pdf-app.vue";
-import ThemeToggling from "@/views/ThemeToggling.vue";
+import ThemeToggling from "@/views/theme-toggling.vue";
 import About from "@/components/about.vue";
-import Sandbox from "@/views/Sandbox.vue";
+import Sandbox from "@/views/sandbox.vue";
 
 Vue.use(Router);